chore: reducer 함수 새로운 파일로 분리
